[win32] peripherals: move implementation to platform folder.
