+++ xbmc/peripherals/bus/PeripheralBusUSB.h
@@ -21,7 +21,7 @@
 
 #if   defined(TARGET_WINDOWS_DESKTOP)
 #define HAVE_PERIPHERAL_BUS_USB 1
-#include "win32/PeripheralBusUSB.h"
+#include "platform/win32/peripherals/PeripheralBusUSB.h"
 #elif defined(TARGET_WINDOWS_STORE)
 #define HAVE_PERIPHERAL_BUS_USB 1
 #include "platform/win10/peripherals/PeripheralBusUSB.h"
