+++ xbmc/platform/win32/peripherals/PeripheralBusUSB.h
